+++ .gitignore
@@ -1,5 +1,4 @@
 .direnv/*
 .envrc
 /target
-/wasm/build
 shell.nix
